requirements.txt
unsloth[colab-new]@git+https://github.com/unslothai/unsloth.git
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
DataProperty==1.1.0
datasets==3.6.0
dill==0.3.8
dnspython==2.8.0
evaluate==0.4.6
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.28
lm_eval==0.4.9.1
lxml==6.0.1
markdown2==2.5.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mbstrdecoder==1.1.4
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numexpr==2.12.1
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pathvalidate==3.3.1
peft==0.17.1
platformdirs==4.4.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pybind11==3.0.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.15.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.38.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlitedict==2.1.0
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.56.1
typepy==1.3.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.4
word2number==1.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
